WIP: More intuitive building/installation of packages
pmbootstrap has already gotten some support for "timestamp based rebuilds", which modifies the logic for when packages should be rebuilt. It doesn't only consider packages outdated with old pkgver/pkgrel combinations, but also packages, where a source file has a newer timestamp, than the built package has.

I've found out, that this can lead to more rebuilds than expected. For example, when you check out the pmbootstrap git repository again into another folder, although you have already built packages. Then all files have the timestamp of the checkout, and the packages will appear to be outdated. While this is not largely a concern now, this will become a problem once we have a binary package repository, because then the packages from the binary repo will always seem to be outdated, if you just freshly checked out the repository.

To combat this, git gets asked if the files from the aport we're looking at are in sync with upstream, or not. Only when the files are *not* in sync with upstream *and* the timestamps of the sources are newer, a rebuild gets triggered from now on.

In case this logic should fail, I've added an option during "pmbootstrap init" where you can enable or disable the "timestamp based rebuilds" option.

In addition to that, this commit also works on fixing #120: packages do not get updated in "pmbootstrap install" after they have been rebuilt. For this to work, we specify all packages explicitly for abuild, instead of letting abuild do the resolving. This feature will also work with the "timestamp based rebuilds". This feature isn't working quite right yet, it's annoying that pmbootstrap displays all dependencies in the install command (especially the redundant so: files), and sometimes it seems to complain about non-existing versions for no reason.

This commit also fixes the working_dir argument in pmb.helpers.run.user, which was simply ignored before.

Finally, I've noticed that the performance got worse by reading Alpine's installed packages list directly instead of calling apk to let it display the version (this change is already in master!). But this change was necessary, again for the "timestamp based rebuilds". Sorry for that, I think this can be resolved in the next commits in this branch.